Allocate destination buffer outside loop for more reliable timing
The "easy" jpegls_decoder::decode call allocates a buffer, fills it with zeros (standard behaviour of std::vector) and copies the decoded pixels into the buffer.
Allocating the buffer outside the loop ensures that only the decoding algorithm is benchmarked, and excludes the overhead how fast std::vector can allocate a buffer.